End of Section 18

+++ client/src/app/messages/messages.component.ts
@@ -1,3 +1,4 @@
+import { ConfirmService } from './../_services/confirm.service';
 import { MessageService } from './../_services/message.service';
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
@@ -17,7 +18,10 @@ export class MessagesComponent implements OnInit {
 	pageSize = 5;
 	loading = false;
 
-	constructor(private messageService: MessageService) {}
+	constructor(
+		private messageService: MessageService,
+		private confirmService: ConfirmService
+	) {}
 
 	ngOnInit(): void {
 		this.loadMessages();
@@ -40,11 +44,17 @@ export class MessagesComponent implements OnInit {
 	}
 
 	deleteMessage(id: number): void {
-		this.messageService.deleteMessage(id).subscribe(() => {
-			this.messages.splice(
-				this.messages.findIndex((m) => m.id === id),
-				1
-			);
-		});
+		this.confirmService
+			.confirm('Confirm delete message', 'This cannot be undone')
+			.subscribe((result) => {
+				if (result) {
+					this.messageService.deleteMessage(id).subscribe(() => {
+						this.messages.splice(
+							this.messages.findIndex((m) => m.id === id),
+							1
+						);
+					});
+				}
+			});
 	}
 }
